Replaced floating based ADPCM decoding routine with fixed point.
git-svn-id: http://svn.purei.org/purei/trunk@731 b36208d7-6611-0410-8bec-b1987f11c4a2

+++ Source/iop/Iop_SpuBase.h
@@ -233,8 +233,8 @@ namespace Iop
 			uint16			m_pitch;
 			float			m_currentTime;
 			float			m_dstTime;
-			float			m_s1;
-			float			m_s2;
+			int32			m_s1;
+			int32			m_s2;
 			bool			m_done;
 			bool			m_nextValid;
 			bool			m_endFlag;
